# backend/server/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "GimmeBeans API ",
        "version": "1.0.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User login",
                "operationId": "99539a8e18179644cbac3165cf994fe5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string"
                                        },
                                        "user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/signup": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User registration",
                "operationId": "e21e6a924e7c86ec87f929adf46803e7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created successfully"
                    }
                }
            }
        },
...
        "/api/addtobasket": {
            "post": {
                "tags": [
                    "Shopping Basket"
                ],
                "summary": "Add product to shopping basket",
                "operationId": "0aa0d975d9c0017b8555b67c09c0022c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user",
                                    "sku",
                                    "qty",
                                    "ground"
                                ],
                                "properties": {
                                    "user": {
                                        "type": "string",
                                        "example": "user@example.com"
                                    },
                                    "sku": {
                                        "type": "string",
                                        "example": "PROD-001"
                                    },
                                    "qty": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "ground": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 2.5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product added successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "boolean"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/basket": {
            "get": {
                "tags": [
                    "Shopping Basket"
                ],
                "summary": "Get user's shopping basket",
                "operationId": "366e0acac0ed7a7e583c4ad04c6252f4",
                "responses": {
                    "200": {
                        "description": "List of basket items",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "user": {
                                                "type": "string"
                                            },
                                            "sku": {
                                                "type": "string"
                                            },
                                            "ground": {
                                                "type": "number"
                                            },
                                            "qty": {
                                                "type": "integer"
                                            },
                                            "unit_price": {
                                                "type": "number"
                                            },
                                            "total_price": {
                                                "type": "number"
                                            },
                                            "image_nobackground": {
                                                "type": "string"
                                            },
                                            "product_name": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}